backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
transformers==4.47.1
torch==2.5.1
openai-whisper==20240930
python-multipart==0.0.20
pydantic==2.10.4
